+++ changelog.md
@@ -1,6 +1,7 @@
 GDB+ 7.6.0
 * address = dbg.write(None, data) allocates memory to store data
 * Improve callback handling for syscall catchpoints
+* Improve structure parsing to interpret types and pointers
 
 GDB+ 7.5.0 **Parse structures**
 * Parse and read c structures into python objects